Revert "https://imagemagick.org/discourse-server/viewtopic.php?f=3&t=35185"
This reverts commit 8c7648a1adf7bba35594074f191affd3ff3263bb.

After the upgrade from 6.9.10.14 to 6.9.10.23, I noticed that the
autopkgtests for the package ruby-mini-magick is failing due to one test
about checking units to show the size of the image in cm and inches.

I could isolate the problem by running the `identify` command on a test
image (rgb.png in the spec/fixtures directory of the ruby-mini-magick
source package).

bug-debian: https://bugs.debian.org/918642

Signed-off-by: Bastien Roucariès <rouca@debian.org>